+++ CHANGES.rst
@@ -1,9 +1,9 @@
 Changes in TestLink-API-Python-client Source Distribution
 =========================================================
 
-TestLink-API-Python-client UNDER DEVELOP v0.5.2 
+TestLink-API-Python-client release notes v0.5.2 (Oct. 2014) 
 -----------------------------------------------------------
-support for future TL 1.9.12 release
+support for TestLink release 1.9.12
 
 implement 1.9.12 new api method - getTestCaseAssignedTester #29
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -30,6 +30,7 @@ examples see `<example/TestLinkExample.py>`_
   
 TestLink-API-Python-client release notes v0.5.1 (Aug. 2014) 
 ------------------------------------------------------------
+support for TestLink release 1.9.11
 
 implement 1.9.11 api change - getLastExecutionResult #27
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~